refactor: pythonify the module layout
