+++ tropos/docx/rubric.py
